requirements.txt
Django==3.2.7
graphene==2.1.9
graphene-django==2.15.0
PyJWT==2.1.0
django-graphql-jwt==0.3.4
paystackapi==2.0.0
python-Levenshtein==0.12.2
